services/sensor-service/requirements.txt
paho-mqtt==1.6.1
PyYAML
requests 
Flask 
# Flask required for actuator REST API 